recurrence_indexing_spin: Fix typo uint->uint8_t

+++ libraries/spncci/recurrence_indexing_spin.h
@@ -581,7 +581,7 @@ class UnitTensorLabelsST
   unsigned long id() const { return bitrep_.to_ulong(); }
 
   constexpr UnitTensorLabelsST(int S0, int T0, int Sbar, int Sbarp, int Tbar, int Tbarp)
-      : bitrep_{static_cast<uint>(
+      : bitrep_{static_cast<uint8_t>(
           64 * S0 + 16 * T0 + 8 * Sbar + 4 * Sbarp + 2 * Tbar + Tbarp
         )}
   {}
